+++ projects/egeo/src/lib/index.ts
@@ -289,6 +289,7 @@ export { StTableRowComponent } from './st-table/shared/st-table-row/st-table-row
 export { StTableCellComponent } from './st-table/shared/st-table-cell/st-table-cell.component';
 export { StTableHeader } from './st-table/shared/table-header.interface';
 export { Order, ORDER_TYPE } from './st-table/shared/order';
+export { StTableIconClasses, StTableFilterIconClasses} from './st-table/st-table.interface';
 
 // Dynamic Table
 export { StDynamicTableModule } from './st-dynamic-table/st-dynamic-table.module';